refactor: create_app structure.
