+++ .env
@@ -1,4 +1,2 @@
 DB_NAME=app
 DB_PASSWORD=my_secret_password
-
-SECRET_KEY=JWT_SECRET_TO_ENCRYPT
