+++ server/models/user.py
@@ -1,4 +1,4 @@
-from extensions import db
+from app.extensions import db
 from datetime import datetime
 
 class User(db.Model):
